Make serialization external
This commit moves the serialization to the caller side. This causes
slightly bigger code at the call site and the interface macros harder
to test (they require deserialization) in exchange for faster
producers.

It breaks the log function in a prepare and commit function.

The prepare function gives back a prebuilt serialized chunked with
room for serializing the arguments on the outside, which is done by
the hacky macro.

The commit function just pushes the memory to the queue after the
arguments have been serialized.