core: Fix very fast init/terminate issues
There was a bug that could make a terminate call after init call
fail, even when running from the same thread.

The init call wasn't waiting for the st_running state, so the
terminate call could hit the first run state.

The error code was reported correctly, so it could be avoided at the
user side, but this commit makes that transitional state invisible
from the outside by moving some of the consumer initialization to the
initialization point.