core: Do termination through a queue command
This fixes another init/terminate timing bug that could potentially
cause message losses. Sending the command through the queue synces
with the message production, so there is no way to terminate before
processing all the messages that happened before the terminate call.

A nice side-effect of this approach is that the consumer code is
simpler to reason about too.