meson: Fix broken C++ tests when ptr compression is enabled
It turns out that on meson files, "c_args" is silently ignored for
C++ projects. One is meant to use "cpp_args" (not to be confused with
makefile's CPPFLAGS, which is to be consumed by the C preprocessor
(CPP)).

This commit fixes many instances of C++ test programs that were
broken because somehow the compiler switches and preprocessor defines
were mistakenly passed to "c_flags". This was visible only when
data compression was enabled.