WARNING: ref values: dealloc on err and filtered severities
Previously to this commit, the documentation stated that when using
dynamic memory on log entries ("logstrref" AND/OR "logmemref" +
"logrefdtor") the user had to manually deallocate the memory he was
passing in presence of an error condition.

This had two problems:

-Filtered out log severities return no error. The user can't know if
that he needs to deallocate the memory and malc wasn't doing it
either. I did a design errata when I forgot to account for this
feature, where returning something like "bl_nothing_to_do" could have
been enough to keep the old behavior.

-It's not convenient for the user to write the deallocation code each
time dynamic memory is used.

This commit adds automatic in-place deallocation of dynamic values
for filtered out severity and error cases. The user doesn't have to
care anymore about the passed memory. Once is passed to malc it will
be destructed by malc, no exceptions or shared responsibilities.

I considered returning "bl_nothing_to_do" for filtered out severities
instead, but this would break every call on already written (correct)
code that checks errors.

I have been lucky this time, as if I had been given the choice when
implementing this from scratch I'd have chosen the solution
implemented on this commit anyways, as it has benefits:

-It's more convenient and cleaner for client code. This is always to
be prefered.

-Only breaks code using "logstrref" AND/OR "logmemref" + "logrefdtor",
which I suspect is a minor and seldomly used feature given the
intended use case of this logger.

-The destruction code is only generated for calls using dynamic
values.

It has drawbacks too:

-It's not very convenient and clean for the library code, as it's
implemented on the already insane macro-abusing code autogeneration
part. I had no alternative, as it's the only way to accomplish what
is accomplished here by using only C.

-This feature, depending on the use case, can forcefully make unwanted
deallocations to happen on the fast-path. It might be a problem or not
depending on what the destructor (logrefdtor) is doing.

This commit add the WARNING label to catch attention, as this is an
API change the major version is increased. The ABI is left unchanged.