cpp: Add wrapping for the destinations.
This commit adds some C++ wrapping to use the existing destinations
in a more "object oriented" way.

From a C++ point of view, there is still a lot of error checking
boilerplate on intialization because the code isn't exception based.
Maybe I will encapsulate the error handling and allow throwing for
code outside of the fast-path. I'm afraid this is going in the
direction of adding more and more templated policy-based code...