malc: Don't require the explicit termination sequence.
The termination sequence on malc, errors ignored, is something like
this:

-Calling "malc_terminate" from some thread.
-Running the "malc_run_consume_task" function until it returns
 "bl_preconditions" to empty malc's queue.
-Destroying.

This was annoying, especially on C++ code that could.

What was preventing "malc_destroy" to run this sequence was that
"malc_run_consume_task" wasn't thread safe.

This commit introduces a mutex on "malc_run_consume_task" and
implements the shutdown sequence described above on "malc_destroy".

The mutex is not going to degrade performance, as it's going to be:

-Mostly uncontended.
-Masked by the I/O operations going on on the consumer function.

This commit bumps "base_library" to get mutex wrappers:

1115fe6 thread: add mutex wrappers