cpp wrapper: Move to namespace malcpp.
This commit moves every C++ related object to the malcpp namespace.

Unfortunately C artifacts still leak to the root namespace.

This might be improved by minimizing inclusions to malc.h and all the
destinations by placing them on an own static library.

If this is done, Every C struct and function to be shared between the
C and CPP versions could be on a common file and the file included
from the CPP version inside the namespace.

Unfortunately this comes at the cost of segmenting the interface,
so it's not done yet.