cpp: Move malc calls inside a wrapper.
This commit is a step towards having a full C++ non-macro based
implementation.

Its aim is to move as much C functions and structs away from the
global namespace as possible.

This is done by moving all C++ wrapping code that calls the orginal
C code to a small wrapping library. On the small wrapping library
both the C and the C++ versions are glued, so this doesn't happen
in client code.

Common structs are moved to <include/common.h>. This file now allows
a double inclusion, one from C and the other one from C++, which
places the structs inside a namespace.

The public headers of the C++ implementation only see the namespaced
structs. On the C implementation they are non-namespaced. The wrapper
library compiled here sees both and forwards the calls via casting
the structs from namespaced to not namespaced.

All the logging variadic functions/macros, validations,
transformations, in place-serialization, etc. are to be done. This
commit leaves the C++ implementatation in an incomplete state that
needs to include <malc/malc.h>, so the C implementation hiding
effects of this commit are not visible yet, they will be when the
producer side of malc is reimplemented in C++.