malc: entry parser improvements
The documentation for the entry string on the readme file is
improved on this commit. This is part of the efforts of setting up a
specification on what the compile-time format parsing of the C++11
has to do.

This commit improvements some things found on the entry parser while
studying it:

-Adds some comments.

-Adds the '#' flag.

-Protects the integer printf format string generator from being
 accidentally passed multiple widths.

-Protects the integer printf format string generator from being
 accidentally passed the '.' symbol multiple times.

-Removes the 'N' nibbles specifier from floats, as there was no way
 to output floats a hex anyways. I guess I thought that 'a'/'A' were
 just outputting in a format they don't when I wrote the parser.