Change malc's logging macros error from parameter to return
This is another interface breaking change.

When the C++11 wrapper is done the error code will be returned from
an anonymous lambda.

The C version is restricted to use the compound statement as
expression extension on the logging macro. AFAIK this is restricted
to GCC and CLANG only. So from now on I assume that only people using
the C version will be using GCC or CLANG. This leaves, on most cases,
Windows users using the C++ version.