c++11: object serialization: pass data through a table
When serializing malc_obj, previously there next data was passed "on
the wire":

-get_data function.
-destroy function.
-size of the object.

This commit defines a table containing all this data. Now the type
implementor has to provide a pointer to a table and keep it alive for
the program duration (e.g. by making it static const). This has the
next advantages:

-optimized serialization: 2 pointers and a char become 1 pointer.
-allowing passing extra static data, the get_data function is
 modified to take a pointer to the passed table, so implementations
 can use the table for passing extra static data (next commits).