c++11: serialization: create rvalue wrapper
On the serialization types, previously there was a hidden convention
of passing rvalues as std::moved copies of the type and lvalues as
pointers.

This commit uses an explicit boolean parameter, so things seem less
magic.

This is at the expense of clarity. The code get more complicated,
mainly because there is now a lot of move constructor and assignment
operator boilerplate.