drop dynamic pointer compression option
It turns out that this option wasn't ideal, as when pointers have
trailing zeros they use to do it in a constant per-platform amount.

This commit makes the pointer compression to have a fixed amount of
bytes instead of being detected at run time for each passed pointer.

The "ptr_bytes_cut_count" meson option is introduced.