c++: fix allocator passing
The allocators were passed by value before, so it wasn't possible to
use "container_of" like techniques on them when provided externally.

malc C++ now stores a pointer and a table, so when no allocator is
passed it uses a pointer pointing to its owned allocator. When an
allocator is passed the own allocator (by value) remains unused.